# docs/swagger.yaml
definitions:
  entity.Cluster:
    properties:
      Status:
        type: string
      api_server:
        type: string
      createdAt:
        type: string
      id:
        type: string
      kube_config:
        type: string
      name:
        type: string
      token:
        type: string
      type:
        type: string
      updatedAt:
        type: string
      user_id:
        type: string
      version:
        type: string
    type: object
  entity.ClusterPage:
    properties:
      items:
        items:
          $ref: '#/definitions/entity.Cluster'
        type: array
      total:
        type: integer
    type: object
  entity.HelmInstance:
    properties:
      chart_name:
        type: string
      cluster_id:
        type: string
      create_namespace:
        type: boolean
      namespace:
        type: string
      release_name:
        type: string
      repo_id:
        type: string
      values_yaml:
        type: string
    type: object
  entity.HelmRepositry:
    properties:
      ca_file:
        type: string
      cert_file:
        type: string
      createdAt:
        type: string
      id:
        type: string
      insecure_skip_tls_verify:
        type: boolean
      is_active:
        type: boolean
      key_file:
        type: string
      name:
        type: string
      password:
        type: string
      updatedAt:
        type: string
      url:
        type: string
      username:
        type: string
    type: object
...
  entity.LoginProfile:
    properties:
      captcha_id:
        type: string
      code:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
...
  entity.Profile:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
      user:
        $ref: '#/definitions/entity.User'
    type: object
...
  entity.User:
    properties:
      createdAt:
        type: string
      id:
        type: string
      name:
        type: string
      password:
        type: string
      role:
        type: integer
      roles:
        items:
          type: string
        type: array
      type:
        type: string
      updatedAt:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /api/v1/auth/login/:
    post:
      consumes:
      - application/json
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.LoginProfile'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Profile'
      tags:
      - 登录
  /api/v1/auth/logout/:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      responses: {}
      tags:
      - 登出
  /api/v1/auth/refresh/:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Profile'
      tags:
      - 刷新token
  /api/v1/auth/register/:
    post:
      consumes:
      - application/json
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.User'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Profile'
      tags:
      - 注册
  /api/v1/cluster:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户ID
        in: query
        name: user_id
        required: true
        type: string
      - description: 集群名称
        in: query
        name: name
        required: true
        type: string
      responses:
        "200":
          description: OK
      tags:
      - 根据用户ID和集群名称删除集群
    get:
      consumes:
      - application/json
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: 集群名
        in: query
        name: cluster_name
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Cluster'
      tags:
      - 根据集群名获取集群
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.Cluster'
      responses:
        "200":
          description: OK
      tags:
      - 根据集群名获取集群
  /api/v1/cluster/helm/application:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.HelmInstance'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.HelmInstance'
      tags:
      - 列出集群已经安装的helm应用
  /api/v1/cluster/helm/repository:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.HelmInstance'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.HelmInstance'
      tags:
      - 添加仓库到集群
  /api/v1/clusters:
    get:
      consumes:
      - application/json
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: 页码
        in: query
        name: page
        required: true
        type: string
      - description: 长度
        in: query
        name: size
        required: true
        type: string
      - description: 用户ID
        in: query
        name: user_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.ClusterPage'
      tags:
      - 根据用户获取集群
  /api/v1/clusters/:cluster_id:
    get:
      consumes:
      - application/json
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: 集群ID
        in: query
        name: cluster_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Cluster'
      tags:
      - 根据集群ID获取集群
  /api/v1/helm/repository:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户名
        in: query
        name: name
        required: true
        type: string
      responses:
        "200":
          description: OK
      tags:
      - 删除仓库
    patch:
      consumes:
      - application/json
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.HelmRepositry'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.HelmRepositry'
      tags:
      - 更新仓库
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.HelmRepositry'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.HelmRepositry'
      tags:
      - 添加仓库
...
swagger: "2.0"